(BKR-1507) New Method for vcenter_creds
Supplement load_config to handle credentials for multiple vcenter instances
so that we aren't limited to a single set of vcenter credentials.

This is a more aptly named method for the purpose which can be used in
place of the existing method for the remainder of those modules that
grab vsphere creds - once that is done, the load_config method can be
dropped.